+++ crates/onetaskgraph-github-projects/CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.9](https://github.com/nickderobertis/onetaskgraph/compare/onetaskgraph-github-projects-v0.2.8...onetaskgraph-github-projects-v0.2.9) - 2026-08-28
+
+### Added
+
+- *(github-projects)* store many projects on one board as issues and their sub-issues ([#56](https://github.com/nickderobertis/onetaskgraph/pull/56))
+
+## [0.3.0](https://github.com/nickderobertis/onetaskgraph/compare/onetaskgraph-github-projects-v0.2.8...onetaskgraph-github-projects-v0.3.0) - 2026-08-28
+
+### Added
+
+- *(github-projects)* store many projects on one board as issues and their sub-issues ([#56](https://github.com/nickderobertis/onetaskgraph/pull/56))
+
 ## [0.2.8](https://github.com/nickderobertis/onetaskgraph/compare/onetaskgraph-github-projects-v0.2.7...onetaskgraph-github-projects-v0.2.8) - 2026-08-28
 
 ### Fixed